Modified to work with hippocampus data
Now able to strip "eye" from session directories named sessioneye.

Now changes directories to the channel directories so the qsub, output, and error files will be created in the channel directories. This way, the progress of the sorting can be checked more easily.

Increased the request for CPU time in the qsub files to make sure long jobs will be able to complete.

Now uses username to path to hpctmp2.